Bug 1640839 - Add some helpers for enums to calculate the size and use it in EnumeratedArray. r=glandium,jgilbert,media-playback-reviewers,padenot
Differential Revision: https://phabricator.services.mozilla.com/D201334

UltraBlame original commit: 24a9665c6ced903ff65e94e323e0a921783b1e12

+++ dom/base/FlushType.h
@@ -40,7 +40,7 @@ enum class FlushType : uint8_t {
 
 
 
-const EnumeratedArray<FlushType, const char*, FlushType::Count>
+const EnumeratedArray<FlushType, const char*, size_t(FlushType::Count)>
     kFlushTypeNames = {
   "",
   "Event",
