+++ dom/media/ipc/RemoteDecoderManagerChild.cpp
@@ -51,7 +51,7 @@ using namespace gfx;
 
 StaticMutex sLaunchMutex;
 static EnumeratedArray<RemoteDecodeIn, StaticRefPtr<GenericNonExclusivePromise>,
-                       RemoteDecodeIn::SENTINEL>
+                       size_t(RemoteDecodeIn::SENTINEL)>
     sLaunchPromises MOZ_GUARDED_BY(sLaunchMutex);
 
 
@@ -62,7 +62,7 @@ static StaticDataMutex<StaticRefPtr<nsIThread>>
 
 
 static EnumeratedArray<RemoteDecodeIn, StaticRefPtr<RemoteDecoderManagerChild>,
-                       RemoteDecodeIn::SENTINEL>
+                       size_t(RemoteDecodeIn::SENTINEL)>
     sRemoteDecoderManagerChildForProcesses;
 
 static StaticAutoPtr<nsTArray<RefPtr<Runnable>>> sRecreateTasks;
@@ -71,7 +71,7 @@ static StaticAutoPtr<nsTArray<RefPtr<Runnable>>> sRecreateTasks;
 
 StaticMutex sProcessSupportedMutex;
 static EnumeratedArray<RemoteDecodeIn, Maybe<media::MediaCodecsSupported>,
-                       RemoteDecodeIn::SENTINEL>
+                       size_t(RemoteDecodeIn::SENTINEL)>
     sProcessSupported MOZ_GUARDED_BY(sProcessSupportedMutex);
 
 class ShutdownObserver final : public nsIObserver {
